Robot Programmer in Scratch
I developed this educational game after noticing that the students at the college where I teach were struggling to learn programming. The root of the problem was their difficulty in grasping the fundamental concept of programming: essentially, giving instructions in a logical sequence. This game aims to make learning these basics more engaging and intuitive.